We now have working support for openblas + FFTW.
There is a very strange bug/issue, where it seems that if OpenBlas was
not compiled with USE_OPENMP=1, then for some odd reason, FFTW seems
to not use threading if the final executable or library was linked
with open blas. For now, I have recompiled open-blas with openmp
support and this is actually faster than MKL now.