Add support for ATLAS blas. Revert to mkl-not single shared lib.
You can now specify openblas or atlas with configure flag.

MKL is now linked with several libs, rather than the single libkl_rt,
because I think it is dangerous that the behaivior depends on
catastrophyically depending on calls to mkl_set_interface_layer,
and I would rather users dont have to worry about this, though maybe
my problem was that my c-driver program did not link to my shared lib,
but compiled it in directly?

Note that the Intel MKL link advisor does not work for calling the
library from matlab and python: we need to add extra libs for that to
work, for some bizzar reason. The internet is full of people with
failed numpy builds for this exact problem...