Change the l1qc_newton integration test to matrix transforms.
This exposed a "bug" in l1qc_newton, because the matrix transforms require
that the output argument be initialized, where as the dct transforms do not.

Fix this by changing all the allocations to l1c_calloc_double and friends.
This is overkill, but doesnt add much time, and I dont have to worry about
it now.