Completely generalize the linesearch.
We now have two linesearches, one for functions of x and one
for (x,u). Each line search to works with an arbitrary function point,
and these are now in their own file.

Add small unit tests for the linesearches based on theory with small quadratic functionals.

Cleanup docs and CI for new linesearch.