Build tests for dct_tv and fix bugs they reveal
So far, only dct2 path tested.

Must be careful with pointer arithmetic when we only have, e.g.,
alp_h>0, because then effectively, the portion of u associated with
alp_v is empty, and we only allocate 2*m space for u. So if we are not
careful, we end up accessing u[2m:3m].