The returned sizes of l1c_mex_utils should be l1c_int, not size_t.
For now anyway, because sizeof(size_t)=8, and sizeof(l1c_int)=4, we can't cast
the pointers safely. Not sure what to do with this, because really, we should be
using size_t for all the index stuff, but many of the blas libs take a standard
int, which is how l1c_int is defined.