Add mode parameters for basis pursuit mode(analsys/synthesis) and dct1/dct2.
The reason for dct1/dct2 is that I would rather return the proper shape,
than rely on the caller reshaping the input to get dct1 or dct2.

Also, that model breaks when we want to do TV, because then we need to have
a 2D input, even if we want dct1.

The purpose of BpMode is to adjust how ax_funs gets constructed, rather than
relying on the caller setting  Mx=NULL, or something funky like that. It is better
I think to make the unused operators the identity.

We also add Wtx and Wz, for analysis mode.
Maybe there is a better way. This seems like its a little out of hand.