Remove memory leaks from python interface.
When building an output array from data we allocated, we cant free our own
memory, because numpy just uses a reference to it, but can't decref it
properly either. This creates a memory leak. So the point of the new
function is to create an output array, and copy our data into it, so
we can free our own memory.